Examied different segmentation proc., read below
Segmentation based on the temporal variance
(i.e. the variance for each pixel along the time-axis) combined with
Otsu thresholding does not work properly (procedure similar to what has
been suggested in Puybareau et al. 2015 and 2016).
I also tried Otsu thresholding for the power dissipated in the
CBF-bandwidth, which does not work either.