Added RMSD to measure quality of fit
The root mean squared deviation between the fit function and the power
spectral density gets now determined to measure the quality of the fit.
The RMSD (normalized by the peak to allow for comparability) may serve
as a measure for how well the CBF-bandwidth can be determined.
Furthermore, a treeview widget has been added to display the results
determined during the automated analysis.